requirements.txt
git+git://github.com/Kerem-Sami-Coop/caishen_stonks
requests==2.25.1
PyYAML==5.4.1
dash==1.20.0
pandas==1.2.4
colorlover==0.3.0
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
bootstrap-flask
dash_bootstrap_components
gunicorn
